+++ src/Deeplearning.Sample/Utils/OxyPlotHelper.cs
@@ -74,9 +74,9 @@ namespace Deeplearning.Sample.Utils
             for (int j = 0; j < vectorCount; j++)
             {
                
-                    float x = matrix[0, j];
+                    double x = matrix[0, j];
 
-                    float y = matrix[1, j];
+                    double y = matrix[1, j];
 
                     yield return new ScatterPoint(x, y);
